requirements.txt
asgiref==3.7.2
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
Markdown==3.4.4
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
